+++ fetch/full_system/her/gym/envs/toy_text/guessing_game.py
@@ -1,8 +1,7 @@
-import numpy as np
-
 import gym
 from gym import spaces
 from gym.utils import seeding
+import numpy as np
 
 
 class GuessingGame(gym.Env):
@@ -41,8 +40,7 @@ class GuessingGame(gym.Env):
         self.range = 1000  # Randomly selected number is within +/- this value
         self.bounds = 10000
 
-        self.action_space = spaces.Box(low=np.array([-self.bounds]), high=np.array([self.bounds]),
-                                       dtype=np.float32)
+        self.action_space = spaces.Box(low=np.array([-self.bounds]), high=np.array([self.bounds]))
         self.observation_space = spaces.Discrete(4)
 
         self.number = 0
