+++ fetch/full_system/her/gym/envs/toy_text/hotter_colder.py
@@ -1,8 +1,7 @@
-import numpy as np
-
 import gym
 from gym import spaces
 from gym.utils import seeding
+import numpy as np
 
 
 class HotterColder(gym.Env):
@@ -26,8 +25,7 @@ class HotterColder(gym.Env):
         self.range = 1000  # +/- value the randomly select number can be between
         self.bounds = 2000  # Action space bounds
 
-        self.action_space = spaces.Box(low=np.array([-self.bounds]), high=np.array([self.bounds]),
-                                       dtype=np.float32)
+        self.action_space = spaces.Box(low=np.array([-self.bounds]), high=np.array([self.bounds]))
         self.observation_space = spaces.Discrete(4)
 
         self.number = 0
